--- node/config_opts.go
@@ -1,15 +1,0 @@
-package node
-
-// WithConfig sets the entire custom config.
-func WithConfig(custom *Config) Option {
-	return func(sets *settings) {
-		sets.cfg = custom
-	}
-}
-
-// WithMutualPeers sets the `MutualPeers` field in the config.
-func WithMutualPeers(addrs []string) Option {
-	return func(sets *settings) {
-		sets.cfg.P2P.MutualPeers = addrs
-	}
-}
