+++ node/core/config.go
@@ -1,0 +1,14 @@
+package core
+
+// Config combines all configuration fields for managing the relationship with a Core node.
+type Config struct {
+	IP       string
+	RPCPort  string
+	GRPCPort string
+}
+
+// DefaultConfig returns default configuration for managing the
+// node's connection to a Celestia-Core endpoint.
+func DefaultConfig() Config {
+	return Config{}
+}
