+++ node/node_light_test.go
@@ -4,25 +4,27 @@ import (
 	"crypto/rand"
 	"testing"
 
+	"github.com/celestiaorg/celestia-node/node/p2p"
+
 	"github.com/libp2p/go-libp2p-core/crypto"
 	mocknet "github.com/libp2p/go-libp2p/p2p/net/mock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/celestiaorg/celestia-node/node/node"
+	nodebuilder "github.com/celestiaorg/celestia-node/node/node"
 	"github.com/celestiaorg/celestia-node/params"
 )
 
 func TestNewLightWithP2PKey(t *testing.T) {
 	key, _, err := crypto.GenerateEd25519Key(rand.Reader)
 	require.NoError(t, err)
-	node := TestNode(t, node.Light, WithP2PKey(key))
+	node := TestNode(t, nodebuilder.Light, WithP2POption(p2p.WithP2PKey(key)))
 	assert.True(t, node.Host.ID().MatchesPrivateKey(key))
 }
 
 func TestNewLightWithHost(t *testing.T) {
 	nw, _ := mocknet.WithNPeers(1)
-	node := TestNode(t, node.Light, WithHost(nw.Hosts()[0]))
+	node := TestNode(t, nodebuilder.Light, WithP2POption(p2p.WithHost(nw.Hosts()[0])))
 	assert.Equal(t, nw.Peers()[0], node.Host.ID())
 }
 
@@ -31,13 +33,13 @@ func TestLight_WithMutualPeers(t *testing.T) {
 		"/ip6/100:0:114b:abc5:e13a:c32f:7a9e:f00a/tcp/2121/p2p/12D3KooWSRqDfpLsQxpyUhLC9oXHD2WuZ2y5FWzDri7LT4Dw9fSi",
 		"/ip4/192.168.1.10/tcp/2121/p2p/12D3KooWSRqDfpLsQxpyUhLC9oXHD2WuZ2y5FWzDri7LT4Dw9fSi",
 	}
-	node := TestNode(t, node.Light, WithMutualPeers(peers))
+	node := TestNode(t, nodebuilder.Light, WithP2POption(p2p.WithMutualPeers(peers)))
 	require.NotNil(t, node)
 	assert.Equal(t, node.Config.P2P.MutualPeers, peers)
 }
 
 func TestLight_WithNetwork(t *testing.T) {
-	node := TestNode(t, node.Light)
+	node := TestNode(t, nodebuilder.Light)
 	require.NotNil(t, node)
 	assert.Equal(t, params.Private, node.Network)
 }
